@@ -2,7 +2,7 @@ name: example
 description: A simple command-line application.
 
 environment:
-  sdk: '>=2.8.1 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   local_firebase_auth:
