@@ -1,16 +1,16 @@
 name: local_firebase_auth
 description: An experimental dart package that mocks (emulates) the firebase_auth plugin for quick and offline development.
-version: 0.3.0+1
+version: 1.0.0-nullsafety.0
 homepage: https://github.com/Akora-IngDKB/local_firebase_auth.git
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
-  hive: ^1.4.4
-  meta: ^1.2.3
-  random_string: ^2.1.0
+  hive: ^2.0.1
+  meta: ^1.3.0
+  random_string: ^2.2.0-nullsafety
 
 dev_dependencies:
-  pedantic: ^1.9.0
-  test: ^1.15.4
+  pedantic: ^1.11.0
+  test: ^1.16.8
